Move cli package to internal/cli
