+++ internal/cli/destroy.go
