+++ internal/cli/info.go
