+++ internal/cli/registry_add.go
