+++ internal/cli/registry_delete.go
