+++ internal/cli/registry_help.go
