+++ internal/cli/registry_list.go
