+++ internal/cli/stop.go
